Modyfied implicit casts off.

+++ ft_atoi.c
@@ -29,12 +29,12 @@ int	ft_atoi(const char *str)
 	}
 	while (*str && *str >= '0' && *str <= '9')
 	{
-		res = res * 10 + *str - '0';
+		res = res * 10 + (unsigned long int)*str - '0';
 		if (res > 9223372036854775807 && sign < 0)
 			return (0);
 		if (res > 9223372036854775807 && sign > 0)
 			return (-1);
 		str++;
 	}
-	return ((int)(sign * res));
+	return (sign * (int)res);
 }
